requirements.txt
PyQt6
PyMuPDF
python-docx
Pillow
pydub
moviepy
tqdm
pdf2docx
pyinstaller
requests
semver